Solve issue with KEYSTREAM info

+++ os/lib/cbor/cbor.h
@@ -54,7 +54,7 @@ int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, size_t bytes_len);
 
 int cbor_put_map(uint8_t **buffer, uint8_t elements);
 
-int cbor_put_unsigned(uint8_t **buffer, uint8_t value);
+int cbor_put_unsigned(uint8_t **buffer, size_t value);
 
 int cbor_put_negative(uint8_t **buffer, int64_t value);
